move contents of test_auth.rs to auth.rs
