+++ contracts/axelar-gateway/src/lib.rs
@@ -15,9 +15,6 @@ compile_error!("'testutils' feature is not supported on 'wasm' target");
 #[cfg(any(test, feature = "testutils"))]
 pub mod testutils;
 
-#[cfg(test)]
-mod test_auth;
-
 // Allows using std (and its macros) in test modules
 #[cfg(test)]
 #[macro_use]
